pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "text-generation-pipeline"
version = "0.1.0"
description = "Efficient Large-Scale Synthetic Text Generation Pipeline"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "transformers>=4.30.0",
    "vllm>=0.2.0",
    "wandb>=0.15.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "pyarrow>=12.0.0",
    "pandas>=2.0.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.0.1",
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py" 